+++ src/components/Modals/Changelog/index.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import whatsNew from "spcr-whats-new";
 import { RELEASE_CHANGELOG, MARKETPLACE_VERSION } from "../../../constants";
+import { getReadmeHTML } from "../../../logic/Utils";
 
 const fetchReleaseBody = async () => {
   return fetch(RELEASE_CHANGELOG)
@@ -8,38 +9,13 @@ const fetchReleaseBody = async () => {
     .then(async (result) => {
       // If API returns no error message then get release body
       const body = !result.message ? result.body as string : null;
-      if (body) return await getReadmeHTML(body);
+      if (body) return await getReadmeHTML(body, "spicetify", "spicetify-marketplace");
       return body;
     })
     .catch((err) => {
       console.error(err);
       return null;
     });
-};
-
-const getReadmeHTML = async (readmeText: string) => {
-  try {
-    const postBody = {
-      text: readmeText,
-      context: `spicetify/spicetify-marketplace`,
-      mode: "gfm",
-    };
-
-    const readmeHtmlRes = await fetch("https://api.github.com/markdown", {
-      method: "POST",
-      body: JSON.stringify(postBody),
-    });
-    if (!readmeHtmlRes.ok)
-      throw Spicetify.showNotification(
-        `Error parsing README (HTTP ${readmeHtmlRes.status})`,
-      );
-
-    const readmeHtml = await readmeHtmlRes.text();
-    return readmeHtml;
-  } catch (err) {
-    console.error(err);
-    return null;
-  }
 };
 
 const Changelog = async () => {
